requirements.txt
google-generativeai>=0.3.0
flask>=2.0.0
pandas>=1.3.0
scikit-learn>=1.0.0
tensorflow>=2.10.0
nltk>=3.6.0
transformers>=4.20.0
torch>=1.10.0
numpy>=1.20.0
fastapi>=0.95.0
uvicorn>=0.22.0
pydantic>=1.10.0
python-dotenv>=0.20.0
requests>=2.28.0
joblib>=1.1.0 